+++ src/components/Signup/index.tsx
@@ -3,7 +3,7 @@ import { Formik, Form, FormikHelpers } from 'formik';
 import { loginSchema } from '../../schemas/loginSchema';
 
 import Logo from '../common/Logo';
-import Input from '../common/FormComponents/Input';
+import Input from '../common/FormComponents/temp.tsx';
 import Select from '../common/FormComponents/select';
 import CheckBox from '../common/FormComponents/checkbox';
 import PrimaryButton from '../common/Button/primary';
